task: moved search data to content search comp
